+++ src/Adapters/Http/Boss4D.Adapters.Http.pas
@@ -10,6 +10,8 @@ type
   TBoss4DHttpNativeAdapter = class(TInterfacedObject, IBoss4DHttpClient)
   public
     function Get(const AURL: string; out AResponse: string): Integer;
+    function PostJson(const AURL, ABody: string;
+      out AResponse: string): Integer;
   end;
 
 implementation
@@ -51,6 +53,40 @@ begin
       LStringStream.Free;
     end;
   finally
+    LClient.Free;
+  end;
+end;
+
+function TBoss4DHttpNativeAdapter.PostJson(const AURL, ABody: string;
+  out AResponse: string): Integer;
+var
+  LClient: THTTPClient;
+  LResponse: IHTTPResponse;
+  LBody, LOutput: TStringStream;
+begin
+  AResponse := '';
+  LClient := THTTPClient.Create;
+  LBody := TStringStream.Create(ABody, TEncoding.UTF8);
+  LOutput := TStringStream.Create('', TEncoding.UTF8);
+  try
+    LClient.UserAgent := 'Boss4D/1.3';
+    LClient.ContentType := 'application/json';
+    LClient.ConnectionTimeout := 10000;
+    LClient.ResponseTimeout := 15000;
+    try
+      LResponse := LClient.Post(AURL, LBody, LOutput);
+      Result := LResponse.StatusCode;
+      AResponse := LOutput.DataString;
+    except
+      on E: Exception do
+      begin
+        Result := 500;
+        AResponse := E.Message;
+      end;
+    end;
+  finally
+    LOutput.Free;
+    LBody.Free;
     LClient.Free;
   end;
 end;