+++ src/Core/Ports/Boss4D.Core.Ports.pas
@@ -85,6 +85,8 @@ type
   IBoss4DHttpClient = interface
     ['{69527D56-F14E-43D4-A746-2D7227D60005}']
     function Get(const AURL: string; out AResponse: string): Integer;
+    function PostJson(const AURL, ABody: string;
+      out AResponse: string): Integer;
   end;
 
   { Contrato para compilacao de dependencias Delphi e search paths }